feat: abstract atom creator for relations to another file
This can be generally used across the app
